code change
- update forward movement magnitude to take into account 3 decimal place
- update targetAngle, targetDistance, avgDist to double
- update actionBuffer code to be able to receive END00001 which we will not be sending STOP0000 when dequeue this

+++ Core/Src/motor.c
@@ -19,8 +19,8 @@ extern uint8_t isMoving;
 extern uint8_t isAngle;
 
 extern double totalAngle;
-extern int targetAngle;
-extern int targetDistance;
+extern double targetAngle;
+extern double targetDistance;
 
 extern uint8_t OLED_row1[20], OLED_row4[20];
